90% Working version of sharesheets for React-native 0.33
